refactor: merge pathfinder-config into the binary
