--- crates/config/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "pathfinder-config"
-version = "0.1.0"
-edition = "2021"
-license = "MIT OR Apache-2.0"
-rust-version = "1.62"
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-clap = { version = "3.1.6", features = ["env"] }
-enum-iterator = "0.7.0"
-lazy_static = "1.4.0"
-reqwest = { version = "0.11.4", features = ["json"] }
-serde = { version = "1.0.130", features = ["derive"] }
-toml = "0.5.8"
-
-[build-dependencies]
-pathfinder-build = { path = "../build" }
