--- crates/config/build.rs
@@ -1,3 +1,0 @@
-fn main() {
-    pathfinder_build::setup_vergen();
-}
